test: split config flow validation tests
